"move function to registry"
